+++ src/store/product/product.action.js
@@ -10,7 +10,7 @@ export const getProductsAction = () => {
     dispatch({type: Types.PRODUCT_LOADING, status: true});
     try {
       const result = await getProductsService();
-      dispatch({type: Types.PRODUCT_LIST, data: result.data.data});
+      dispatch({type: Types.PRODUCT_LIST, data: result.data.data[0].data});
     } catch (error) {}
   };
 };
@@ -30,7 +30,7 @@ export const getProductsSearchAction = name => {
     dispatch({type: Types.PRODUCT_LOADING, status: true});
     try {
       const result = await getProductsSearchService(name);
-      dispatch({type: Types.PRODUCT_LIST, data: result.data.data});
+      dispatch({type: Types.PRODUCT_LIST, data: result.data.data[0].data});
     } catch (error) {}
   };
 };
