+++ src/util/helpers/format.js
@@ -1,6 +1,6 @@
 export function formatPriceBr(data) {
   return Number(data).toLocaleString('pt-br', {
     style: 'currency',
-    currency: 'BRL',
-  });
+    currency: 'BRL'
+  })
 }
